+++ assets/src/components/parts/janus-text-flow/Markup.tsx
@@ -104,7 +104,10 @@ const Markup: React.FC<any> = ({
     // empty elements in HTML don't stay in the flow
     // add a non breaking space instead of nothing
 
-    processedText = processedText.length < 2 ? '\u00a0' : processedText.replace(/ \s/g, '\u00a0 ');
+    processedText =
+      processedText.length < 2 && !processedText.trim()
+        ? '\u00a0'
+        : processedText.replace(/ \s/g, '\u00a0 ');
   } else if (processedText.length !== processedText.trim().length) {
     // check if text has leading and trailing spaces.
     //handling the leading blank spacecs in the span
@@ -287,6 +290,13 @@ const Markup: React.FC<any> = ({
           {processedText}
           {children}
         </code>
+      );
+    case 'blockquote':
+      return (
+        <blockquote ref={el} key={key} className={customCssClass} style={renderStyles}>
+          {processedText}
+          {children}
+        </blockquote>
       );
     case 'ol':
       return (