+++ CHANGELOG.md
@@ -3,8 +3,11 @@
 ## Unreleased
 
 ### Bug Fixes
+- Fix an issue with Apply Update button and move content updates to async worker
 
 ### Enhancements
+
+- New Admin landing page
 
 ## 0.14.0 (2021-10-13)
 
@@ -32,7 +35,6 @@
 - Add ability to remix materials from multiple projects
 - Fix insert content popup in page editor
 - Add blackboard LTI 1.3 configuration instructions
-
 
 ### Release Notes